+++ backend/src/main/java/com/zzol/sizzang/review/dto/request/ReviewGetReq.java
@@ -7,6 +7,6 @@ import lombok.Data;
 public class ReviewGetReq {
 
     private int mkCode;
-    private int stCode;
-    private int reCode;
+    private Long stCode;
+    private Long reCode;
 }
